backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
supabase==1.0.3
pyjwt==2.8.0
loguru==0.4.6
pydantic==2.7.0
sqlalchemy==2.0.20